requirements.txt
flask
flask-jwt-extended
psycopg2-binary
bcrypt
python-dotenv
pandas
flask-cors
